Developement of web page for user; uses CSS, HTML and libraries and session  implementation.
1. Make a home (or index) page. With at least three links in them. 
Home 
Registration 
login if you click on registration, a registration form will open that includes captcha and some 
a minimal field like username, password, email address, name along with some a text 
field to enter a captcha. 
After clicking submit button your program has to verify the captcha entered as well as 
duplicate entries with the database. 
If a duplicate entry is there show some appropriate message. 
Otherwise, make an entry in your database. 
On clicking the login username, the password field should be shown along with a link to the registration page. After clicking submit button you have to check username and 
password against your database. 
If correct, login to user 
Otherwise, display an appropriate message. 
Note:- for registering as well as for login logout session implementation should be 
there. 
User is free to choose any library and function to include captcha in their page